ai-buddy/requirements.txt
# Core Streamlit
streamlit>=1.45.1

# Database connections
pymongo>=4.0.0
faiss-cpu>=1.7.4

# AI/ML APIs
openai>=1.0.0

# Vector embeddings & processing
sentence-transformers>=2.2.2
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# PDF processing
PyPDF2>=3.0.0
pymupdf>=1.23.0
pdfplumber>=0.9.0
pypdf>=3.0.0

# Text processing
langchain>=0.1.0
langchain-community>=0.0.10
tiktoken>=0.5.0

# Utilities
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
Pillow>=10.0.0

# Optional: Hugging Face
transformers>=4.30.0
accelerate>=0.20.0
onnxruntime>=1.16.0

audio-recorder-streamlit>=0.0.8